lnwallet: add hardware wallet
Add a second wallet type that supports a hardware (or hybrid) wallet.
lnd communicates with the wallet using an RPC interface. A mock RPC
wallet server is provided for testing.

This also required introducing an abstraction for the blockchain
backends (bitcoind, btc, neutrino) that can be shared between
the two wallet implementations.